moved code for download docs in functions and remove ALL functions from view to "func.py"
